Start removing LIRICAL genomics options.
