+++ maxodiff-html/src/main/java/org/monarchinitiative/maxodiff/html/config/LiricalAutoConfiguration.java
@@ -1,10 +1,6 @@
 package org.monarchinitiative.maxodiff.html.config;
 
-import org.monarchinitiative.lirical.configuration.impl.BundledBackgroundVariantFrequencyServiceFactory;
-import org.monarchinitiative.lirical.core.model.GenomeBuild;
 import org.monarchinitiative.lirical.core.service.PhenotypeService;
-import org.monarchinitiative.lirical.core.service.VariantMetadataServiceFactory;
-import org.monarchinitiative.lirical.exomiser_db_adapter.ExomiserMvStoreMetadataServiceFactory;
 import org.monarchinitiative.maxodiff.config.MaxodiffDataResolver;
 import org.monarchinitiative.maxodiff.lirical.LiricalDifferentialDiagnosisEngineConfigurer;
 import org.monarchinitiative.maxodiff.lirical.MaxodiffLiricalAnalysisRunner;
@@ -18,10 +14,6 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.HashMap;
-import java.util.Map;
 
 
 @Configuration
@@ -56,47 +48,47 @@ public class LiricalAutoConfiguration {
 
     @Bean
     public MaxodiffLiricalAnalysisRunner liricalAnalysisRunner(PhenotypeService liricalPhenotypeService) {
-        BundledBackgroundVariantFrequencyServiceFactory factory = BundledBackgroundVariantFrequencyServiceFactory.getInstance();
+//        BundledBackgroundVariantFrequencyServiceFactory factory = BundledBackgroundVariantFrequencyServiceFactory.getInstance();
         int parallelism = 4; // TODO: extract into a property
-        return MaxodiffLiricalAnalysisRunnerImpl.of(liricalPhenotypeService, factory, parallelism);
+        return MaxodiffLiricalAnalysisRunnerImpl.of(liricalPhenotypeService, parallelism);
     }
 
-    @Bean
-    public VariantMetadataServiceFactory variantMetadataServiceFactory(LiricalProperties liricalProperties) {
-        Map<GenomeBuild, Path> exomiserDbPaths = new HashMap<>();
+//    @Bean
+//    public VariantMetadataServiceFactory variantMetadataServiceFactory(LiricalProperties liricalProperties) {
+//        Map<GenomeBuild, Path> exomiserDbPaths = new HashMap<>();
+//
+//        // Check HG19
+//        checkExomiserDbPath(
+//                GenomeBuild.HG19,
+//                liricalProperties.getExomiserHg19Path(),
+//                exomiserDbPaths
+//        );
+//
+//        // Check HG38
+//        checkExomiserDbPath(
+//                GenomeBuild.HG38,
+//                liricalProperties.getExomiserHg38Path(),
+//                exomiserDbPaths
+//        );
+//
+//        return ExomiserMvStoreMetadataServiceFactory.of(exomiserDbPaths);
+//    }
 
-        // Check HG19
-        checkExomiserDbPath(
-                GenomeBuild.HG19,
-                liricalProperties.getExomiserHg19Path(),
-                exomiserDbPaths
-        );
-
-        // Check HG38
-        checkExomiserDbPath(
-                GenomeBuild.HG38,
-                liricalProperties.getExomiserHg38Path(),
-                exomiserDbPaths
-        );
-
-        return ExomiserMvStoreMetadataServiceFactory.of(exomiserDbPaths);
-    }
-
-    private static void checkExomiserDbPath(
-            GenomeBuild genomeBuild,
-            String pathString,
-            Map<GenomeBuild, Path> exomiserDbPaths
-    ) {
-        if (pathString != null) {
-            Path path = Path.of(pathString);
-            if (Files.isRegularFile(path)) {
-                LOGGER.debug("Using Exomiser database at {} for {}", path.toAbsolutePath(), genomeBuild);
-                exomiserDbPaths.put(genomeBuild, path);
-            } else {
-                throw new RuntimeException("Not a file: %s".formatted(path.toAbsolutePath()));
-            }
-        }
-    }
+//    private static void checkExomiserDbPath(
+//            GenomeBuild genomeBuild,
+//            String pathString,
+//            Map<GenomeBuild, Path> exomiserDbPaths
+//    ) {
+//        if (pathString != null) {
+//            Path path = Path.of(pathString);
+//            if (Files.isRegularFile(path)) {
+//                LOGGER.debug("Using Exomiser database at {} for {}", path.toAbsolutePath(), genomeBuild);
+//                exomiserDbPaths.put(genomeBuild, path);
+//            } else {
+//                throw new RuntimeException("Not a file: %s".formatted(path.toAbsolutePath()));
+//            }
+//        }
+//    }
 
     @Bean
     public LiricalDifferentialDiagnosisEngineConfigurer liricalDifferentialDiagnosisEngineConfigurer(
