+++ maxodiff-html/src/main/resources/templates/liricalAnalysis.html
@@ -39,14 +39,14 @@
             </tr>
             </thead>
             <tbody>
-            <tr>
-                <td>Genome Build:</td><td th:text="${options == null} ? '' : *{genomeBuild}"></td>
-                <td>Transcript Database:</td><td th:text="${options == null} ? '' : *{transcriptDatabase}"></td>
-            </tr>
-            <tr>
-                <td>Pathogenicity Threshold:</td><td th:text="${options == null} ? '' : *{variantDeleteriousnessThreshold}"></td>
-                <td>Default Variant Background Frequency:</td><td th:text="${options == null} ? '' : *{defaultVariantBackgroundFrequency}"></td>
-            </tr>
+<!--            <tr>-->
+<!--                <td>Genome Build:</td><td th:text="${options == null} ? '' : *{genomeBuild}"></td>-->
+<!--                <td>Transcript Database:</td><td th:text="${options == null} ? '' : *{transcriptDatabase}"></td>-->
+<!--            </tr>-->
+<!--            <tr>-->
+<!--                <td>Pathogenicity Threshold:</td><td th:text="${options == null} ? '' : *{variantDeleteriousnessThreshold}"></td>-->
+<!--                <td>Default Variant Background Frequency:</td><td th:text="${options == null} ? '' : *{defaultVariantBackgroundFrequency}"></td>-->
+<!--            </tr>-->
             <tr>
                 <td>Strict:</td><td th:text="${options == null} ? '' : *{useStrictPenalties}"></td>
                 <td>Global Mode:</td><td th:text="${options == null} ? '' : *{useGlobal}"></td>
